Move utils methods into one class
