+++ core/src/main/java/novoda/lib/sqliteprovider/util/DBUtils.java
@@ -1,6 +1,7 @@
 
 package novoda.lib.sqliteprovider.util;
 
+import android.content.ContentValues;
 import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
 
@@ -13,6 +14,7 @@ import java.util.Map;
 import java.util.Map.Entry;
 
 import novoda.lib.sqliteprovider.sqlite.IDatabaseMetaInfo.SQLiteType;
+import novoda.rest.database.SQLiteTableCreator;
 
 public final class DBUtils {
 
@@ -117,6 +119,18 @@ public final class DBUtils {
     }
 
 
+    public static String getSQLiteVersion2() {
+        final Cursor cursor = SQLiteDatabase.openOrCreateDatabase(":memory:", null).rawQuery(
+                "select sqlite_version() AS sqlite_version", null);
+        StringBuilder sqliteVersion = new StringBuilder();
+        while (cursor.moveToNext()) {
+            sqliteVersion.append(cursor.getString(0));
+        }
+        cursor.close();
+        return sqliteVersion.toString();
+    }
+
+
     /**
      * Use {@link #getUniqueConstraints(SQLiteDatabase, String)}
      */
@@ -139,7 +153,6 @@ public final class DBUtils {
         return constrains;
     }
 
-
     public static List<Constraint> getUniqueConstraints(SQLiteDatabase db, String table) {
         List<Constraint> constraints = new ArrayList<Constraint>();
         final Cursor indexCursor = db.rawQuery(String.format(PRGAMA_INDEX_LIST, table), null);
@@ -159,5 +172,50 @@ public final class DBUtils {
         }
         indexCursor.close();
         return constraints;
+    }
+
+    public static String contentValuestoTableCreate(ContentValues values, String table) {
+        StringBuffer buf = new StringBuffer("CREATE TABLE ").append(table).append(" (");
+        for (Entry<String, Object> entry : values.valueSet()) {
+            buf.append(entry.getKey()).append(" TEXT").append(", ");
+        }
+        buf.delete(buf.length() - 2, buf.length());
+        buf.append(");");
+        return buf.toString();
+    }
+
+    public static String getCreateStatement(SQLiteTableCreator creator) {
+
+        String primaryKey = creator.getPrimaryKey();
+        novoda.rest.database.SQLiteType primaryKeyType;
+        boolean shouldAutoincrement;
+        if (primaryKey == null) {
+            primaryKey = "_id";
+            primaryKeyType = novoda.rest.database.SQLiteType.INTEGER;
+            shouldAutoincrement = true;
+        } else {
+            primaryKeyType = creator.getType(primaryKey);
+            shouldAutoincrement = creator.shouldPKAutoIncrement();
+        }
+
+        StringBuilder sql = new StringBuilder();
+        sql.append("CREATE TABLE IF NOT EXISTS ").append("\"" + creator.getTableName() + "\"")
+        .append(" (").append(primaryKey).append(" ").append(primaryKeyType.name())
+        .append(" PRIMARY KEY").append(((shouldAutoincrement) ? " AUTOINCREMENT " : " "));
+
+        for (String f : creator.getTableFields()) {
+            if (f.equals(primaryKey)) {
+                continue;
+            }
+            sql.append(", ").append(f).append(" ").append(creator.getType(f).name());
+            sql.append(creator.isNullAllowed(f) ? "" : " NOT NULL");
+
+            sql.append(creator.isUnique(f) ? " UNIQUE" : "");
+            sql.append((creator.onConflict(f) != null && creator.isUnique(f)) ? " ON CONFLICT "
+                    + creator.onConflict(f) : "");
+        }
+
+        sql.append(");");
+        return sql.toString();
     }
 }
